requirements.txt
# Core ML and Data Processing
pyspark==3.3.0
tensorflow==2.10.0
scikit-learn==1.1.3
pandas==1.5.1
numpy==1.23.4
pyarrow==10.0.1
fastparquet==0.8.3
imbalanced-learn==0.10.1

# Visualization
matplotlib==3.6.2
seaborn==0.12.1

# Development and Notebooks
jupyter==1.0.0

# MLOps and Experiment Tracking
mlflow==2.0.1
pyyaml==6.0

# Production Dependencies
gunicorn==20.1.0
flask==2.2.3
redis==4.5.1
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.3

# Monitoring and Logging
wandb==0.15.0
psutil==5.9.5

# Testing
pytest==7.3.1
pytest-cov==4.1.0
